fix: streaming availability transformer. Update logs
Logging needed to be switched from console.log to console.error to be more percise on the desired purpose. Streaming availability was so convoluted it made it hard to troubleshoot what was failing. At the end of the day I messed up the hash map method I was using to create the table data so I just revised the whole thing. Clearer to understand now.